Updating parallel script to call functions in neat directly instead of using subscripts
