+++ config_template/template_neat_config.yml
@@ -140,43 +140,35 @@ min_mutations: .
 # type: bool | required = no | default = false
 overwrite_output: .
 
-# Top-level output directory for splits, per-chunk outputs, and stitched results.
-# Relative paths are interpreted against the CURRENT WORKING DIRECTORY.
-# If omitted (or set to .), it defaults to: <cwd>/<config_stem>_parallel
-# type = string | required: no
-outdir: .
-
-# Location (prefix, no extension) for stitched outputs.
-# If relative, it is resolved under outdir (i.e., <outdir>/<final_prefix>*).
-# Default is "stitched/final".
-# type = string | required: no | default = stitched/final
-final_prefix: .
+# Controls whether parallel mode is activated. The following parameters all are used in parallel mode, but not
+# otherwise.
+# type = bool | required: no | default = false
+parallel_mode: .
 
 # How to split the input reference for parallelization
+# If parallel_mode = false, this option has no effect.
 # type = string | required: no | default = contig | values: contig, size
 by: .
 
 # Target chunk size if by = size (overlap = read_len * 2).
-# Default is 500000 when by = size.
+# Default is 500000 when by = size. Not used for by = contig.
 # type = int | required: no | default = 500000 (when by=size)
 size: .
 
-# Maximum number of concurrent NEAT jobs
-# type = int | required: no | default = (CPU count)
-jobs: .
-
-# Command used to launch the simulator (CLI mode)
-# type = string | required: no | default = "neat read-simulator"
-neat_cmd: .
-
-# Path to samtools (binary name if on PATH)
-# type = string | required: no | default = samtools
-samtools: .
+# Maximum number of concurrent NEAT jobs (threads or hyperthreads) to run.
+# Note that assigning a value higher than your available number of threads
+# could result in performance issues in the computer running the job.
+# If parallel_mode = false, this option has no effect.
+# type = int | required: no | default = all available.
+threads: .
 
 # Delete the 'splits' directory after stitching completes
+# Note if parallel_mode = false, this option has no effect.
 # type = bool | required: no | default = false
 cleanup_splits: .
 
-# Reuse existing files in 'splits' and skip the split step
-# type = bool | required: no | default = false
+# Reuse existing files in 'splits' and skip the split step. This
+# option should be a path to the directory containing the splits.
+# Note if parallel_mode = None, this option has no effect.
+# type = str | required: no | default = None
 reuse_splits: .
