--- neat/parallel_read_simulator/__init__.py
@@ -1,5 +1,0 @@
-"""
-Load modules needed for other parts of the program
-"""
-from .parallelize import parallelize_main
-__all__ = ["parallelize_main"]
