+++ neat/read_simulator/__init__.py
@@ -2,3 +2,4 @@
 Modules to generate reads
 """
 from .runner import *
+from .parallel_runner import main
