+++ neat/read_simulator/utils/split_inputs.py
@@ -15,6 +15,10 @@ import logging
 
 __all__ = ["main"]
 
+from numpy.lib.twodim_base import fliplr
+
+from neat.read_simulator.utils import Options
+
 _LOG = logging.getLogger(__name__)
 
 
@@ -42,10 +46,8 @@ def print_stderr(msg: str, *, exit_: bool = False) -> None:
     if exit_:
         sys.exit(1)
 
-
 def disk_bytes_free(path: Path) -> int:
     return shutil.disk_usage(path).free
-
 
 # FASTA helpers
 def parse_fasta(path: Path) -> Iterator[SimpleRecord]:
@@ -70,12 +72,10 @@ def parse_fasta(path: Path) -> Iterator[SimpleRecord]:
 
 def write_fasta(rec: SimpleRecord, out_fa: Path, width: int = 60) -> None:
     """Write a sequence record to a FASTA file, wrapping lines at a given width."""
-    out_fa.parent.mkdir(parents=True, exist_ok=True)
     with out_fa.open("w") as fh:
         fh.write(f">{rec.id}\n")
         for chunk in wrap(rec.seq, width):
             fh.write(chunk + "\n")
-
 
 # Chunk generators
 def chunk_record(record: SimpleRecord, chunk_len: int, overlap: int) -> Iterator[Tuple[SimpleRecord, int]]:
@@ -95,7 +95,6 @@ def chunk_record(record: SimpleRecord, chunk_len: int, overlap: int) -> Iterator
         start = end - overlap
         chunk_id += 1
 
-
 def write_config(template_cfg: dict, out_fa: Path, out_yml: Path) -> None:
     """Write a per‑fragment configuration derived from the template config."""
     cfg_copy = template_cfg.copy()
@@ -104,73 +103,41 @@ def write_config(template_cfg: dict, out_fa: Path, out_yml: Path) -> None:
         cfg_copy["name"] = out_fa.stem
     out_yml.write_text(yaml.safe_dump(cfg_copy, sort_keys=False))
 
-
-# Argument parsing
-def parse_args(argv: List[str] | None = None) -> argparse.Namespace:
-    p = argparse.ArgumentParser(
-        description="Split a FASTA and NEAT config into per-contig or fixed-size chunks ready for neat read-simulator.",
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-    )
-    p.add_argument("fasta", type=Path, help="Input multi-contig FASTA")
-    p.add_argument("config", type=Path, help="Original NEAT YAML/YML config")
-    p.add_argument("--outdir", type=Path, required=True, help="Directory to write split FASTAs + configs")
-    p.add_argument(
-        "--by",
-        choices=["contig", "size"],
-        default="contig",
-        help="Split whole contigs (default) or fixed-size chunks",
-    )
-    p.add_argument(
-        "--size",
-        type=int,
-        default=500000,
-        help="Target chunk size when --by size is chosen",
-    )
-    return p.parse_args(argv)
-
-
-def main(argv: List[str] | None = None) -> None:
+def main(options: Options) -> list:
     """Perform the splitting of a FASTA and NEAT configuration."""
-    args = parse_args(argv)
+    overlap = options.read_len * 2
 
-    template_cfg = yaml.safe_load(args.config.read_text())
-    read_len = int(template_cfg.get("read_len", 150))
-    overlap = read_len * 2
-
-    outdir = args.outdir
-    outdir.mkdir(parents=True, exist_ok=True)
-
-    approx_out_bytes = int(args.fasta.stat().st_size * 1.1)
-    if disk_bytes_free(args.outdir) < approx_out_bytes:
+    approx_out_bytes = int(options.fasta.stat().st_size * 1.1)
+    if disk_bytes_free(options.output) < approx_out_bytes:
         print_stderr(
-            f"Not enough free space in {args.outdir} (need about {approx_out_bytes/1e9:.2f} GB)",
+            f"Not enough free space in {options.output} (need about {approx_out_bytes/1e9:.2f} GB)",
             exit_=True,
         )
 
     idx = 1
     pad = 6  # zero-pad width for global indices
 
-    mode = args.by
     written = 0
-
-    for rec in parse_fasta(args.fasta):
-        if mode == "contig":
+    fasta_files = []
+    for rec in parse_fasta(options.fasta):
+        if options.mode == "contig":
             stem = f"{idx:0{pad}d}__{rec.id}"
-            fa = outdir / f"{stem}.fa"
-            yml = outdir / f"{stem}.yaml"
+            fa = options.splits_dir / f"{stem}.fa"
             write_fasta(rec, fa)
-            write_config(template_cfg, fa, yml)
+            fasta_files.append(fa)
             idx += 1
             written += 1
         else:
-            for subrec, chunk_id in chunk_record(rec, args.size, overlap):
+            for subrec, chunk_id in chunk_record(rec, options.size, overlap):
                 stem = f"{idx:0{pad}d}__{subrec.id}"
-                fa = outdir / f"{stem}.fa"
-                yml = outdir / f"{stem}.yaml"
+                fa = options.splits_dir / f"{stem}.fa"
                 write_fasta(subrec, fa)
-                write_config(template_cfg, fa, yml)
+                fasta_files.append(fa)
                 idx += 1
                 written += 1
 
     # Report success via the logger instead of printing to stderr
-    _LOG.info(f"Generated {written} FASTA/YAML pair(s) in {outdir}")
+    _LOG.info(f"Generated {written} FASTAs in {options.splits_dir}")
+    files_written_string = "\n\t".join(fasta_files)
+    _LOG.debug(f'Splits files written: \n\t{files_written_string}')
+    return files_written_string
