+++ neat/read_simulator/utils/stitch_outputs.py
